Update build.sh to check before updating gocbcore v10
Reviewed-on: http://review.couchbase.org/c/query/+/163900
Tested-by: Donald Haggart <donald.haggart@couchbase.com>

+++ build.sh
@@ -73,7 +73,19 @@ DevStandaloneSetup() {
        then
            (cd ..; rm -rf gocbcore/v9; git clone -b v9.1.7 https://github.com/couchbase/gocbcore.git gocbcore/v9)
        fi
-       (cd ..; rm -rf gocbcore/v10; git clone -b v10.0.3 https://github.com/couchbase/gocbcore.git gocbcore/v10)
+    # gocbcore points to master; gocbcore/v10 points to 10.0.3
+       if [[ -d ../gocbcore/v10 ]]
+       then
+           cd ../gocbcore/v10
+           C=`git describe --tags|grep -c "v10.0.3"`
+           cd -
+       else
+           C=0
+       fi
+       if [[ $C -eq 0 ]]
+       then
+           (cd ..; rm -rf gocbcore/v10; git clone -b v10.0.3 https://github.com/couchbase/gocbcore.git gocbcore/v10)
+       fi
     # bleve version
        if [[ ! -d ../../blevesearch/bleve/v2 ]]; then
            (cd ../../blevesearch; git clone -b v2.2.1 http://github.com/blevesearch/bleve.git bleve/v2)
